requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
idna==2.10
lxml==4.6.3
pkg-resources==0.0.0
python-dotenv==0.18.0
requests==2.25.1
soupsieve==2.2.1
urllib3==1.26.6
